pyproject.toml
[project]
name = "maleficer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "esper>=3.3",
    "pyyaml>=6.0.2",
    "ruff>=0.9.7",
    "tcod>=16.2.3",
]
